+++ CHANGELOG.md
@@ -15,4 +15,4 @@
 
 - Upgrade JlrsLedger to v0.1.0
 
-- `Ledger.LEDGER_API_VERSION` is now 2: `Ledger.PoisonError` has been removed, and `Ledger.unborrow_shared` returns `true` if there are no remaining shared borrows or `false` otherwise. 
+- `Ledger.LEDGER_API_VERSION` is now 2: `Ledger.PoisonError` has been removed, and `Ledger.unborrow_shared` returns `true` if there are no remaining shared borrows or `false` otherwise.